feat: export types in separate modules
