+++ _internal_types.ts
@@ -1,0 +1,3 @@
+export type MutableTuple<T extends readonly unknown[]> =
+  // deno-lint-ignore no-explicit-any
+  ((...args: T) => any) extends ((...args: infer U) => any) ? U : never;
