+++ errors/mod.ts
@@ -1,3 +1,11 @@
+/**
+ * Provides error classes for [streamtest][].
+ *
+ * [streamtest]: https://jsr.io/@milly/streamtest
+ *
+ * @module
+ */
+
 export * from "./leaking_async_ops_error.ts";
 export * from "./max_ticks_exceeded_error.ts";
 export * from "./operation_not_permitted_error.ts";
