+++ mod.ts
@@ -1,2 +1,13 @@
+/**
+ * Provides helper functions that make it easier to test [Streams API][]
+ * with various scenarios and assertions.
+ *
+ * [streamtest]: https://jsr.io/@milly/streamtest
+ * [Streams API]: https://developer.mozilla.org/docs/Web/API/Streams_API
+ *
+ * @module
+ */
+
 export * from "./errors/mod.ts";
 export * from "./test_stream.ts";
+export * from "./types.ts";
